refactor : MoveRuleByMinimumScore가 NumberGenrator를 구현하도록 함
