+++ src/test/java/racingcar/domain/MoveRuleByMinimumScoreTest.java
@@ -1,6 +1,8 @@
 package racingcar.domain;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -34,5 +36,14 @@ class MoveRuleByMinimumScoreTest {
         //when
         //then
         assertThat(moveRule.tryMove(3)).isZero();
+    }
+
+    @RepeatedTest(value = 20)
+    @DisplayName("0 ~ 9 사이의 숫자가 나오면 성공한다.")
+    void RandomNumberGeneratorTest() {
+        //given
+        MoveRule numberGenerator = new MoveRuleByMinimumScore();
+        //when,then
+        Assertions.assertThat(numberGenerator.generate()).isBetween(0, 9);
     }
 }
